finished all part 4 exercises

+++ part4/blog_list/backend/controllers/blog.js
@@ -1,38 +1,54 @@
 const blogRouter = require('express').Router()
+const userExtractor = require('../utils/middleware').userExtractor
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/Blog')
+const User = require('../models/User')
 
 blogRouter.get('/', async (request, response) => {
-    const posts = await Blog.find({})
+    const posts = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(posts)
   })
   
-  blogRouter.post('/', async (request, response) => {
-    const blog = request.body
-    const newPost = new Blog ({
-      title: blog.title,
-      author: blog.author,
-      url: blog.url,
-      likes: blog.likes === undefined ? 0 : blog.likes
-    })
-    const post = await newPost.save()
-    response.json(post)
+blogRouter.post('/', userExtractor, async (request, response) => {
+  const blog = request.body
+  const user = request.user
+
+  const newPost = new Blog({
+    title: blog.title,
+    author: blog.author,
+    url: blog.url,
+    likes: blog.likes === undefined ? 0 : blog.likes,
+    user: user._id
   })
 
-  blogRouter.put('/:id', async (request, response) => {
-    const body = request.body
-    const newPost = {
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes
-    }
-    const updatedPost = await Blog.findByIdAndUpdate(request.params.id, newPost, { runValidators: true, new: true, context: 'query' })
-    response.json(updatedPost.toJSON())
-  })
+  const savedPost = await newPost.save()
+  user.blogs = user.blogs.concat(newPost._id)
+  await user.save()
+  response.json(savedPost)
+})
 
-  blogRouter.delete('/:id', async (request, response) => {
+blogRouter.put('/:id', async (request, response) => {
+  const body = request.body
+  const newPost = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes
+  }
+  const updatedPost = await Blog.findByIdAndUpdate(request.params.id, newPost, { runValidators: true, new: true, context: 'query' })
+  response.json(updatedPost.toJSON())
+})
+
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user._id.toString()) {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
-  })
+  }
+  else {
+    response.status(401).json({ error: "you are not authorized to delete this blog post" })
+  }
+})
 
-  module.exports = blogRouter
+module.exports = blogRouter
